Rename 'info' to 'status' and move plugins list into 'plugin-list'
The list of plugins can be inconvenient in the output of 'status'
command if it is too long.
This patch adds a new command 'plugin-list' that list all plugins along
with their version and their state (active, inactive, uninstalled, or
version-mismatch if the database version is different from ini version)
'status' command now displays only the number of installed and
uninstalled plugins.